+++ libtulip/src/main/java/com/tajmoti/libtulip/misc/ListUtil.kt
@@ -11,6 +11,15 @@ inline fun <T> List<T?>.takeIfNoneNull(): List<T>? {
     }
 }
 
+inline fun <T> List<T?>.takeNotNullIfAny(): List<T>? {
+    val notNull = filterNotNull()
+    return if (notNull.isEmpty()) {
+        null
+    } else {
+        notNull
+    }
+}
+
 inline fun <A, B> Pair<A?, B?>.takeIfNoneNull(): Pair<A, B>? {
     val (a, b) = this
     a?.let { b?.let { return Pair(a, b) } }