+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Changed
 - Increase thumb size on seekbar in video player
+- Search results now appear as they load instead of waiting for the last one before they are shown
 
 ## [0.9.1] - 2021-09-25
 ### Changed